+++ DawnLib.Interfaces/src/Interfaces/IDawnPlayerAnimationEvents.cs
@@ -1,0 +1,11 @@
+using InjectionLibrary.Attributes;
+
+[assembly: RequiresInjections]
+
+namespace Dawn.Interfaces;
+
+[InjectInterface(typeof(PlayerAnimationEvents))]
+public interface IDawnPlayerAnimationEvents
+{
+    public void PlayCrouchFootstepSound();
+}